+++ src/app/layout.js
@@ -1,5 +1,5 @@
 import LayoutAvatar from "@/components/auth/layout-avatar";
-import Header from "../components/header";
+import Header from "../components/menu";
 import { NextAuthProvider } from "../context/next-auth-provider";
 import "./globals.css";
