--- src/components/header/header.css
@@ -1,105 +1,0 @@
-/* Button icon for navigation menu */
-.button {
-  position: fixed;
-  left: 0;
-  margin: 20px;
-  width: 80px;
-  height: 80px;
-  background-color: #92bfc5;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: none;
-  cursor: pointer;
-  z-index: 10;
-  top: 0;
-}
-
-/* Burger icon within button */
-.button .burger {
-  width: 100%;
-}
-
-.button .burger::after,
-.button .burger::before {
-  content: "";
-  display: block;
-  width: 40%;
-  background-color: black;
-  height: 3px;
-  margin: auto;
-  position: relative;
-  transition: 0.3s;
-}
-
-.button .burger::after {
-  top: -5px;
-}
-
-.button .burger::before {
-  top: 5px;
-}
-
-/* Burger icon animation */
-.button .burgerActive::after {
-  top: -1px;
-  transform: rotate(45deg);
-}
-
-.button .burgerActive::before {
-  top: 1.5px;
-  transform: rotate(-45deg);
-}
-
-/* Media query for screens 800px or less */
-@media (max-width: 980px) {
-
-  .button {
-    width: 60px;
-    height: 60px;
-    margin: 15px;
-  }
-
-  .button .burger::after,
-  .button .burger::before {
-    width: 35%;
-    height: 2px;
-  }
-
-  .button .burger::after {
-    top: -4px;
-  }
-
-  .button .burger::before {
-    top: 4px;
-  }
-}
-
-/* Media query for screens 515px or less */
-@media (max-width: 750px) {
-
-  .button {
-    width: 50px;
-    height: 50px;
-    margin: 10px;
-  }
-
-  .button .burger::after,
-  .button .burger::before {
-    width: 30%;
-    height: 2px;
-  }
-
-  .button .burger::after {
-    top: -3px;
-  }
-
-  .button .burger::before {
-    top: 3px;
-  }
-}
-
-
-
-
