--- src/components/header/nav/headerNav.css
@@ -1,32 +1,0 @@
-/* Navigation panel styles */
-.menu {
-  position: fixed;
-  left: 0;
-  height: 100vh;
-  background-color: #adc1c6;
-  color: white;
-  z-index: 9;
-  top: 0px;
-}
-
-/* Navigation panel content styles */
-.menu .body {
-  padding: 50px;
-  display: flex;
-  justify-content: space-between;
-}
-
-/* Navigation panel link styles */
-.menu .body .nav {
-  display: flex;
-  flex-direction: column;
-  font-size: 56px;
-  gap: 12px;
-  margin-top: 80px;
-}
-
-.menu .body .nav a {
-  text-decoration: none;
-  color: rgb(39, 39, 39);
-  font-weight: 300;
-}
