--- src/components/header/nav/link/headerLink.css
@@ -1,19 +1,0 @@
-.nav-link {
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  color: inherit;
-  margin-bottom: 1rem;
-  font-size: 1.25rem;
-}
-
-.arrow-icon {
-  margin-left: 0.5rem;
-  width: 1rem;
-  height: 1rem;
-  transition: transform 0.3s ease;
-}
-
-.nav-link:hover .arrow-icon {
-  transform: translateX(3px);
-}
