+++ src/components/menu/styles/menu.css
@@ -1,0 +1,103 @@
+/* Styles for menu links */
+.nav-link {
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    color: inherit;
+    margin-bottom: 1rem;
+    font-size: 1.5rem;
+}
+
+.arrow-icon {
+    margin-left: 0.5rem;
+    width: 1rem;
+    height: 1rem;
+    transition: transform 0.3s ease;
+}
+
+.nav-link:hover .arrow-icon {
+    transform: translateX(3px);
+}
+
+/* Styles for menu popout */
+.menu {
+    position: fixed;
+    left: 0;
+    height: 100vh;
+    background-color: #adc1c6;
+    color: white;
+    z-index: 9;
+    top: 0px;
+}
+
+.menu .body {
+    padding: 50px;
+    display: flex;
+    justify-content: space-between;
+}
+
+.menu .body .nav {
+    display: flex;
+    flex-direction: column;
+    font-size: 56px;
+    gap: 12px;
+    margin-top: 80px;
+}
+
+.menu .body .nav a {
+    text-decoration: none;
+    color: rgb(39, 39, 39);
+    font-weight: 300;
+}
+
+/* Styles for menu button */
+.button {
+    position: fixed;
+    left: 0;
+    margin: 20px;
+    width: 80px;
+    height: 80px;
+    background-color: #92bfc5;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: none;
+    cursor: pointer;
+    z-index: 10;
+    top: 0;
+}
+
+.button .burger {
+    width: 100%;
+}
+
+.button .burger::after,
+.button .burger::before {
+    content: "";
+    display: block;
+    width: 40%;
+    background-color: black;
+    height: 3px;
+    margin: auto;
+    position: relative;
+    transition: 0.3s;
+}
+
+.button .burger::after {
+    top: -5px;
+}
+
+.button .burger::before {
+    top: 5px;
+}
+
+.button .burgerActive::after {
+    top: -1px;
+    transform: rotate(45deg);
+}
+
+.button .burgerActive::before {
+    top: 1.5px;
+    transform: rotate(-45deg);
+}
